+++ src/ops_data_store/logging.py
@@ -3,5 +3,5 @@ import logging
 
 def init() -> None:
     """Initialize application logging."""
-    logging.basicConfig(level=logging.DEBUG, format="%(asctime)s - %(name)s - %(levelname)s - %(message)s")
+    logging.basicConfig(level=logging.INFO, format="%(asctime)s - %(name)s - %(levelname)s - %(message)s")
     logging.getLogger("app")